got the edit form showing up, but not yet plugged into db
