+++ libraryapp/templates/books/form.html
@@ -7,35 +7,45 @@
     </head>
     <body>
         <h1>Inventory Book</h1>
-
-        <form action="{% url 'libraryapp:books'  %}" method="post">
+        {% if book.id is not None %}
+            <form action="{% url 'libraryapp:book' book.id %}" method="post">
+        {% else %}
+            <form action="{% url 'libraryapp:books' %}" method="post">
+        {% endif %}
             {% csrf_token %}
             <fieldset>
                 <label for="title">Title: </label>
-                <input id="title" type="text" name="title">
+                <input id="title" type="text" name="title" value="{{ book.title }}">
             </fieldset>
             <fieldset>
                 <label for="author">Author: </label>
-                <input id="author" type="text" name="author">
+                <input id="author" type="text" name="author" value="{{ book.author }}">
             </fieldset>
             <fieldset>
                 <label for="year_published">Year of publication: </label>
-                <input id="year_published" type="text" name="year_published">
+                <input id="year_published" type="text" name="year_published" value="{{ book.year_published }}">
             </fieldset>
             <fieldset>
                 <label for="isbn">ISBN: </label>
-                <input id="isbn" type="text" name="isbn">
+                <input id="isbn" type="text" name="isbn" value="{{ book.isbn }}">
             </fieldset>
             <fieldset>
                 <label for="location">Library: </label>
                 <select id="location" type="text" name="location">
                     {% for library in all_libraries %}
-                    <option value="{{ library.id }}">{{ library.title }}</option>
+                    <option {% if library.id == book.location_id %}selected{% endif %} value="{{library.id}}">
+                    {{ library.title }}</option>
                     {% endfor %}
                 </select>
             </fieldset>
 
-            <input type="submit" value="Add It" />
+            {% if book.id is not None %}
+                <input type="hidden" name="actual_method" value="PUT">
+                <input type="submit" value="Update" />
+            {% else %}
+                <input type="submit" value="Create" />
+            {% endif %}
+
         </form>
     </body>
 </html>
