+++ libraryapp/views/__init__.py
@@ -4,7 +4,7 @@ from .librarians.list import list_librarians
 from .libraries.list import list_libraries
 from .home import home
 from .auth.logout import logout_user
-from .books.form import book_form
+from .books.form import book_form, book_edit_form
 from .libraries.form import library_form
 from .librarians.details import librarian_details
 from .libraries.details import library_details
