+++ libraryapp/views/books/details.py
@@ -3,26 +3,7 @@ from django.urls import reverse
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from ..connection import Connection
-
-def get_book(book_id):
-    with sqlite3.connect(Connection.db_path) as conn:
-        conn.row_factory = sqlite3.Row
-        db_cursor = conn.cursor()
-
-        db_cursor.execute("""
-        SELECT
-            b.id,
-            b.title,
-            b.isbn,
-            b.author,
-            b.year_published,
-            b.librarian_id,
-            b.location_id
-        FROM libraryapp_book b
-        WHERE b.id = ?
-        """, (book_id,))
-
-        return db_cursor.fetchone()
+from ..helpers.get_book import get_book
 
 @login_required
 def book_details(request, book_id):
