+++ libraryapp/views/books/form.py
@@ -5,6 +5,7 @@ from django.contrib.auth.decorators import login_required
 # from libraryapp.models import Library
 # from libraryapp.models import model_factory
 from ..connection import Connection
+from ..helpers.get_book import get_book
 
 def get_libraries():
     with sqlite3.connect(Connection.db_path) as conn:
@@ -27,6 +28,20 @@ def book_form(request):
         libraries = get_libraries()
         template = 'books/form.html'
         context = {
+            'all_libraries': libraries
+        }
+
+        return render(request, template, context)
+
+@login_required
+def book_edit_form(request, book_id):
+    if request.method == 'GET':
+        book = get_book(book_id)
+        libraries = get_libraries()
+
+        template = 'books/form.html'
+        context = {
+            'book': book,
             'all_libraries': libraries
         }
