+++ libraryapp/views/helpers/get_book.py
@@ -1,0 +1,22 @@
+import sqlite3
+from ..connection import Connection
+
+def get_book(book_id):
+    with sqlite3.connect(Connection.db_path) as conn:
+        conn.row_factory = sqlite3.Row
+        db_cursor = conn.cursor()
+
+        db_cursor.execute("""
+        SELECT
+            b.id,
+            b.title,
+            b.isbn,
+            b.author,
+            b.year_published,
+            b.librarian_id,
+            b.location_id
+        FROM libraryapp_book b
+        WHERE b.id = ?
+        """, (book_id,))
+
+        return db_cursor.fetchone()
